SCM - add `displayId` to `SourceControlHistoryItem`
