+++ src/vscode-dts/vscode.proposed.scmHistoryProvider.d.ts
@@ -63,6 +63,7 @@ declare module 'vscode' {
 		readonly id: string;
 		readonly parentIds: string[];
 		readonly message: string;
+		readonly displayId?: string;
 		readonly author?: string;
 		readonly icon?: Uri | { light: Uri; dark: Uri } | ThemeIcon;
 		readonly timestamp?: number;
